// frontend/src/api/generated/model/updateResourceRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CloudOps API
 * Infrastructure resource management, monitoring and asynchronous tasks. Timestamps use UTC ISO-8601. Error responses share ApiError; errors contains field violations.
 * OpenAPI spec version: 0.0.1-SNAPSHOT
 */
import type { ResourceConfig } from "./resourceConfig";
import type { UpdateResourceRequestStatus } from "./updateResourceRequestStatus";
import type { UpdateResourceRequestType } from "./updateResourceRequestType";

export interface UpdateResourceRequest {
  /**
   * @minLength 0
   * @maxLength 100
   */
  name: string;
  type: UpdateResourceRequestType;
  status: UpdateResourceRequestStatus;
  /** @exclusiveMinimum 0 */
  organizationId: number;
  config: ResourceConfig;
}